move add_analysis to mip specific module mip.py
